Use `setup_test_component_platform` helper for datetime entity component tests instead of `hass.components`
